More runs script improvements

+++ scripts/remote-control/evalRun.sh
@@ -24,12 +24,15 @@ echo "Evaluating run of ${TEAM_NAMES[$RUNNING_TEAM]} ($RUNNING_TEAM) on $RUNNING
 
 rsync -rcLv $REMOTE_USER@$SERVER:$LOGDIR $HOME
 
-gunzip -c $HOME/$RESCUE_LOGFILE.gz > $HOME/$RESCUE_LOGFILE.tmp
+MAP_EVALDIR=$HOME/$EVALDIR/$RUNNING_MAP
+if [ ! -d $MAP_EVALDIR/$RUNNING_TEAM ]; then
+    gunzip -c $HOME/$RESCUE_LOGFILE.gz > $HOME/$RESCUE_LOGFILE.tmp
 
-evalLog.sh $HOME/$RESCUE_LOGFILE.tmp $RUNNING_MAP $RUNNING_TEAM
+    evalLog.sh $HOME/$RESCUE_LOGFILE.tmp $RUNNING_MAP $RUNNING_TEAM
 
-rm $HOME/$RESCUE_LOGFILE.tmp
+    rm $HOME/$RESCUE_LOGFILE.tmp
+else
+    echo "Rebuilding summary page for $RUNNING_MAP"
 
-echo "Rebuilding summary page for $RUNNING_MAP"
-
-mapSummary.sh $RUNNING_MAP
+    mapSummary.sh $RUNNING_MAP
+fi
